# openapi.yaml
openapi: 3.0.0
info:
  title: 꼬리콥터 API 명세
  description: 꼬리콥터 백엔드 API 명세서입니다. 초기 온보딩, 산책 기능, 코스 및 기록 관리를 위한 API들을 포함합니다.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: 개발 서버
...
paths:
...
  /walk-records:
    post:
      tags:
        - 산책 기능
      summary: 산책 기록 시작
      requestBody:
        description: 산책 시작 정보
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                petId:
                  type: string
                startTime:
                  type: string
                  format: date-time
                startLatitude:
                  type: number
                startLongitude:
                  type: number
      responses:
        '201':
          description: 산책 기록 생성 성공
  /walk-records/{id}/track:
    patch:
      tags:
        - 산책 기능
      summary: 산책 경로 및 데이터 주기적 업데이트
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: 현재 위치 및 거리 데이터
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                latitude:
                  type: number
                longitude:
                  type: number
                distanceDelta:
                  type: number
      responses:
        '200':
          description: 업데이트 성공
  /walk-records/{id}/status:
    patch:
      tags:
        - 산책 기능
      summary: 산책 상태 일시정지/재개
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: 변경할 산책 상태
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [PAUSED, RESUMED]
      responses:
        '200':
          description: 상태 변경 성공
  /walk-records/{id}/end:
    put:
      tags:
        - 산책 기능
      summary: 산책 기록 최종 종료 및 전체 데이터 저장
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 산책 종료 및 저장 성공
...
  /walk-records/{id}/score:
    put:
      tags:
        - 산책 기능
      summary: 꼬리콥터 점수 저장
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: 꼬리콥터 점수
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tailcopterScore:
                  type: number
      responses:
        '200':
          description: 점수 저장 성공
...
  /walk-records/{id}/details:
    get:
      tags:
        - 코스 및 기록 관리
      summary: 산책 경로 및 마킹 이미지 등 상세 정보 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 산책 상세 정보 반환
  /walk-records/{id}/save:
    post:
      tags:
        - 코스 및 기록 관리
      summary: 산책 기록 최종 저장 확정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 산책 기록 저장 성공